Extract main GraphitiConfig and MCPConfig classes
As a developer, I need the main configuration classes (GraphitiConfig and MCPConfig) in a central server configuration file that orchestrates all other config components.
