+++ mcp_server/src/config/__init__.py
@@ -3,8 +3,13 @@
 from .database_config import Neo4jConfig
 from .embedder_config import GraphitiEmbedderConfig
 from .llm_config import GraphitiLLMConfig
+from .server_config import GraphitiConfig, MCPConfig
 
 __all__ = [
+    # Primary server configuration classes
+    "GraphitiConfig",
+    "MCPConfig",
+    # Component configuration classes
     "GraphitiLLMConfig",
     "GraphitiEmbedderConfig",
     "Neo4jConfig",
